brain/requirements.txt
fastapi
uvicorn
anthropic
groq
prometheus-client
python-dotenv